♻ refactor(#117): LetterTemplateList 컴포넌트 분리
