+++ src/app/letter/template/page.tsx
@@ -6,7 +6,6 @@ import { theme } from "@/styles/theme";
 import NavigatorBar from "@/components/common/NavigatorBar";
 import Button from "@/components/common/Button";
 import { useRouter, useSearchParams } from "next/navigation";
-import Image from "next/image";
 import Letter from "@/components/letter/Letter";
 import { useRecoilValue, useSetRecoilState } from "recoil";
 import {
@@ -14,6 +13,8 @@ import {
   useSsrComplectedState,
 } from "@/recoil/letterStore";
 import Loader, { LoaderContainer } from "@/components/common/Loader";
+import LetterTemplateList from "@/components/letter/LetterTemplateList";
+import { ALL_TEMPLATES } from "@/constants/templates";
 
 const LetterTemplatePage = () => {
   const router = useRouter();
@@ -82,19 +83,11 @@ const LetterTemplatePage = () => {
               />
             </LetterContainer>
           </LetterWrapper>
-          <TemplatesList>
-            {[0, 1, 2, 3, 4, 5, 6, 7, 8, 9].map((item) => (
-              <TemplateImage
-                key={item}
-                src={`/assets/letter/background_${item}.png`}
-                width={70}
-                height={70}
-                alt="편지지"
-                $selected={template === item}
-                onClick={() => hanleChangeTemplate(item)}
-              />
-            ))}
-          </TemplatesList>
+          <LetterTemplateList
+            selectedTemplate={template}
+            onChangeTemplate={hanleChangeTemplate}
+            templates={ALL_TEMPLATES}
+          />
           <Page>
             <Current>{template + 1}</Current>/{totalPage}
           </Page>
@@ -239,69 +232,6 @@ const LetterContainer = styled.div`
     min-width: 178px;
     max-height: 182px;
     min-height: 182px;
-  }
-`;
-
-const TemplatesList = styled.div`
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  gap: 12px;
-  height: 78px;
-  padding-left: 4px;
-  margin-top: 69px;
-  margin-bottom: 14px;
-  overflow-x: scroll;
-
-  ::-webkit-scrollbar {
-    display: none;
-  }
-  -ms-overflow-style: none; /* IE, Edge */
-  scrollbar-width: none; /* Firefox */
-
-  @media (max-height: 740px) {
-    margin-top: 30px;
-    margin-bottom: 5px;
-    ${theme.fonts.body14};
-    gap: 11px;
-  }
-
-  @media (max-height: 680px) {
-    margin-top: 15px;
-    margin-bottom: 5px;
-    ${theme.fonts.body14};
-    gap: 11px;
-  }
-
-  @media (max-height: 628px) {
-    margin-top: 12px;
-    margin-bottom: 5px;
-    ${theme.fonts.body14};
-    gap: 11px;
-  }
-`;
-
-const TemplateImage = styled(Image)<{ $selected: boolean }>`
-  width: 70px;
-  height: 70px;
-  border-radius: 8px;
-  box-sizing: content-box;
-
-  ${({ $selected, theme }) =>
-    $selected &&
-    css`
-      box-shadow: 0 0 0 4px ${theme.colors.sub03};
-    `}
-
-  @media (max-height: 628px) {
-    width: 50px;
-    height: 50px;
-    border-radius: 4px;
-    ${({ $selected, theme }) =>
-      $selected &&
-      css`
-        box-shadow: 0 0 0 2px ${theme.colors.sub03};
-      `}
   }
 `;
